+++ Source/AtCoderAnalyzer/CreateOperators/SubtractEnumerateMember.cs
@@ -3,15 +3,15 @@ using Microsoft.CodeAnalysis.CSharp;
 
 namespace AtCoderAnalyzer.CreateOperators
 {
-    internal class AdditionEnumerateMember : OperatorEnumerateMember
+    internal class SubtractEnumerateMember : OperatorEnumerateMember
     {
-        internal AdditionEnumerateMember(ITypeSymbol typeSymbol) : base(typeSymbol) { }
+        internal SubtractEnumerateMember(ITypeSymbol typeSymbol) : base(typeSymbol) { }
 
         protected override SyntaxKind? GetSyntaxKind(IMethodSymbol symbol)
             => symbol switch
             {
                 { Parameters: { Length: not 2 } } => null,
-                { Name: "Add" } => SyntaxKind.AddExpression,
+                { Name: "Subtract" } => SyntaxKind.SubtractExpression,
                 _ => null,
             };
     }