+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.8.0] - Unreleased
 - **Breaking** Move operator interfaces to AtCoder.Operators namespace
+- **Breaking** Separate ISubtractOperator from IAdditionOperator
 - Update CreateOperatorCodeFixProvider
 
 ## [1.7.0] - 2021-09-14